+++ examples/file-transfer/receiver.rs
@@ -51,7 +51,7 @@ pub fn run() {
                             }
                         };
 
-                        network.send(endpoint, ReceiverMsg::CanReceive(able)).unwrap();
+                        network.send(endpoint, ReceiverMsg::CanReceive(able));
                     }
                     SenderMsg::Chunk(data) => {
                         let transfer = transfers.get_mut(&endpoint).unwrap();