+++ examples/readme_example.rs
@@ -38,7 +38,7 @@ fn main() {
                 NetEvent::Message(endpoint, message) => match message {
                     InputMessage::HelloServer(msg) => {
                         println!("Received: {}", msg);
-                        network.send(endpoint, OutputMessage::HelloClient(msg)).unwrap();
+                        network.send(endpoint, OutputMessage::HelloClient(msg));
                     },
                     //Other input messages here
                 },