+++ examples/udp/server.rs
@@ -25,9 +25,7 @@ pub fn run() {
                 NetEvent::Message(endpoint, message) => match message {
                     Message::Greetings(text) => {
                         println!("Client ({}) says: {}", endpoint.addr(), text);
-                        network
-                            .send(endpoint, Message::Greetings("Hi, I hear you".into()))
-                            .unwrap();
+                        network.send(endpoint, Message::Greetings("Hi, I hear you".into()));
                     }
                 },
                 NetEvent::AddedEndpoint(_) => unreachable!(), // Not be generated for UDP